Fix the build warnings for suspend debug
kernel/power/user_sysfs_private.c: In function ‘tlmm_dump_info’:
kernel/power/user_sysfs_private.c:167:41: warning: format ‘%d’ expects argument of type ‘int’,
but argument 4 has type ‘long int’ [-Wformat=]
   p += snprintf(p, size - (p - buf), "(%d)\n", p - buf);
                                        ~^      ~~~~~~~

Reviewed-on: https://gerrit.mot.com/1188825
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key
Tested-by: Jira Key
Reviewed-on: https://gerrit.mot.com/1209374

+++ kernel/power/user_sysfs_private.c
@@ -164,7 +164,7 @@ int tlmm_dump_info(char *buf, int size, int tlmm_num, int tlmm_count)
 			}
 		}
 		pr_debug("%s(), %d\n", __func__, __LINE__);
-		p += snprintf(p, size - (p - buf), "(%d)\n", p - buf);
+		p += snprintf(p, size - (p - buf), "(%d)\n", (int)(p - buf));
 	}
 	return p - buf;
 }
@@ -225,7 +225,7 @@ int tlmm_before_sleep_dump_info(char *buf, int size)
 			p += tlmm_dump_cfg(p, size - (p - buf),
 				i, cfg, output_val);
 		}
-		p += snprintf(p, size - (p - buf), "(%d)\n", p - buf);
+		p += snprintf(p, size - (p - buf), "(%d)\n", (int)(p - buf));
 	}
 	return p - buf;
 }
@@ -296,7 +296,7 @@ int tlmm_tz_dump_info(char *buf, int size)
 		p += snprintf(p, size - (p - buf), "%d: ", i);
 		p += snprintf(p, size - (p - buf), "%d\n", tz_configs[i]);
 	}
-	p += snprintf(p, size - (p - buf), "(%d)\n", p - buf);
+	p += snprintf(p, size - (p - buf), "(%d)\n", (int)(p - buf));
 	return p - buf;
 }
 
@@ -349,7 +349,7 @@ int tlmm_before_sleep_table_dump_info(char *buf, int size)
 		/* cfg &= ~0x40000000; */
 		p += tlmm_dump_cfg(p, size - (p - buf), i, cfg, output_val);
 	}
-	p += snprintf(p, size - (p - buf), "(%d)\n", p - buf);
+	p += snprintf(p, size - (p - buf), "(%d)\n", (int)(p - buf));
 	return p - buf;
 }
